#1946 Extract nested helper class that represents an annotation mirror into its own file
